GCP/requirements.txt
google-cloud-monitoring
google-cloud-storage
Pillow==10.3.0
